+++ src/db.py
@@ -50,7 +50,7 @@ class creator:
     # Column name with the cross-reference id
     XID = 'Protein'
     # Meta terms of Protein
-    META = ['xref_UniProt_Name','Protein','Gene','Species','Description','Comment_Line','prot_UniProt_Class']
+    META = ['xref_UniProt_Name','Protein','Gene','Species','Length','Description','Comment_Line','prot_UniProt_Class']
     # Xreferences terms of Protein
     XTERMS = [
         ('Ensembl',  [('xref_Ensembl_protId','(ENS\w*P\d+[.]?\d*)'),('xref_Ensembl_transcId','(ENS\w*T\d+[.]?\d*)'),('xref_Ensembl_GeneId','(ENS\w*G\d+[.]?\d*)'),('Protein','\[([^\]]*)\]')]),
@@ -69,49 +69,50 @@ class creator:
     HEADERS = [ h for h in META] + [ h[0] for i in XTERMS for h in i[1] if h[0] != 'Protein' ] + [ h[0] for i in CTERMS for h in i[1] if h[0] != 'xref_UniProt_Acc' ]
     TIME = datetime.datetime.now().strftime("%Y%m")
 
+
     '''
     Creates the databases
     '''
-    def __init__(self, s, o, f=None, d=False):
+    def __init__(self, s, o, f=None):
         
         # assign species
         species = s.lower()
         if species in self.SPECIES_LIST:
             self.species = species
             self.proteome_id = self.SPECIES_LIST[self.species]['proteome']
+            self.scientific = self.SPECIES_LIST[self.species]['scientific']
         else:
             sys.exit( "ERROR: Species parameter has been not found. Try with: "+", ".join(self.SPECIES_LIST.keys()) )
         
-        # create output directory if does not exist
-        self.outdir = o
-        if not os.path.exists(self.outdir):
-            os.makedirs(self.outdir, exist_ok=True)
-            
-        # create temporal file
+        # prepare temporal file
         self.TMP_DIR = os.path.dirname(os.path.abspath(__file__)) +'/../tmp/'+ self.TIME +'/'+ self.species
         os.makedirs(self.TMP_DIR, exist_ok=True)
-        logging.debug(self.TMP_DIR)
+        logging.debug(f"TMP_DIR: {self.TMP_DIR}")
         
-        # download sequences from UniProt
-        self.outfname = species +'_'+ self.TIME +'_'+ f if f else ''
-        self.db_fasta = self.outdir +'/'+ self.outfname +'.fasta'                
-        self._download_fasta_db(self.db_fasta, f)
-        
-        # remove duplicate sequences
-        if d:
-            logging.debug("remove duplicate sequences")
-            self._remove_duplicates(self.db_fasta, self.db_fasta)
-            
-        # create report from fasta file with the UniProt accession!! (key_function)
-        self.db_fasta_seqio = SeqIO.index(self.db_fasta, "fasta", key_function=lambda rec : rec.split("|")[1])
+        # prepare cached directory
+        self.CHD_DIR = os.path.dirname(os.path.abspath(__file__)) +'/../cached/'+ self.species
+        os.makedirs(self.CHD_DIR, exist_ok=True)
+        logging.debug(f"CACHED_DIR: {self.CHD_DIR}")
 
-        # create data files
-        self.db_uniprot = self.TMP_DIR +'/'+ self.outfname +'.uniprot.dat'
-        self.db_corum   = self.TMP_DIR +'/'+ ".".join(os.path.basename( self.URL_CORUM ).split(".")[:-1]) # get the filename from the URL (without 'zip' extension)
-        self.db_panther = self.TMP_DIR +'/'+ self.outfname +'.panther.dat'
+        # prepare the output directory if does not exist
+        self.outdir = os.path.dirname(o)
+        if not os.path.exists(self.outdir):
+            os.makedirs(self.outdir, exist_ok=True)
         
-        # create output files
-        self.outfile = self.outdir +'/'+ self.outfname +'.categories.tsv'
+        # define output file
+        self.outfile = o
+
+        # define output files for "create_sb"
+        self.db_uniprot = self.TMP_DIR +'/uniprot.dat'
+        # self.db_uniprot = self.TMP_DIR +'/test_2656.dat'
+        # self.db_uniprot = self.TMP_DIR +'/test_1033.dat'
+        # self.db_uniprot = self.TMP_DIR +'/test_to_solve_duplication.dat'
+        
+        self.db_fasta = self.TMP_DIR +'/proteins.fasta'
+        self.db_corum   = self.TMP_DIR +'/'+ ".".join(os.path.basename( self.URL_CORUM ).split(".")[:-1]) # get the filename from the URL (without 'zip' extension)
+        self.db_panther = self.TMP_DIR +'/panther.dat'
+        self.cached_dir_kegg = self.CHD_DIR +'/kegg'
+        os.makedirs(self.cached_dir_kegg, exist_ok=True)
 
 
     def _delete_tmp_dir(self, dir):
@@ -121,21 +122,6 @@ class creator:
                 os.remove(os.path.join(dir, f))
             except Exception as e:
                 logging.error(e)
-
-
-    def _download_fasta_db(self, outfile, filt):
-        '''
-        Download the fasta database file
-        '''
-        url = self.URL_UNIPROT +'query=proteome:'+ self.proteome_id        
-        if filt and filt == "sw": # filter by SwissProt
-            url += '&fil=reviewed:yes'            
-        elif filt and filt == "tr": # filter by TrEMBL
-            url += '&fil=reviewed:no'
-        url += '&format=fasta'
-        logging.debug('get '+url)
-        urllib.request.urlretrieve(url, outfile)
-
 
     def _remove_duplicates(self, infile, outfile=None):
         '''
@@ -166,24 +152,55 @@ class creator:
         return None
 
 
-    def download_raw_dbs(self, filt):
+    def download_fasta_dbs(self, filt=None, d=None):
+        '''
+        Download the fasta file
+        '''        
+        # filter by SwissProt (Reviwed) if apply
+        if not os.path.isfile(self.db_uniprot):
+            if filt and filt.startswith("pro"): # filter by proteome
+                url = self.URL_UNIPROT +'query=proteome:'+ self.proteome_id
+            else: # by default filter by organism
+                url = self.URL_UNIPROT +'query=organism:'+ self.scientific.replace(" ",'+')
+            if filt and filt.endswith("sw"): # filter by SwissProt
+                url += '&fil=reviewed:yes'            
+            url += '&format=fasta'
+            logging.info("get "+url+" > "+self.outfile)
+            urllib.request.urlretrieve(url, self.outfile)
+        else:
+            logging.info('cached uniprot')
+            
+        # remove duplicate sequences
+        if d:
+            logging.info("remove duplicate sequences")
+            self._remove_duplicates(self.outfile, self.outfile)
+            
+        
+    def download_raw_dbs(self, filt=None):
         '''
         Download the raw databases
-        '''
-        # delete any temporal file
-        # self._delete_tmp_dir(self.TMP_DIR)
-        
-        # UniProt
-        # filter by SwissProt (Reviewd) if apply
+        '''        
+        # UniProt Fasta
+        if not os.path.isfile(self.db_fasta):
+            url = self.URL_UNIPROT +'query=organism:'+ self.scientific.replace(" ",'+') +'&format=fasta'
+            logging.info("get "+url+" > "+self.db_fasta)
+            urllib.request.urlretrieve(url, self.db_fasta)
+        else:
+            logging.info('cached uniprot fasta')
+        logging.info('create report from fasta file with the UniProt accession')
+        with open(self.db_fasta, 'r') as f:
+            fasta_txt = f.read()
+        # create a dictionary with the protein accession and the value is a dict(comment, protein_length)
+        b = [ a[1:].split('\n') if a.startswith('>') else a.split('\n') for a in fasta_txt.split('\n>') ]
+        self.db_fasta_seqio = dict([ (c[0].split("|")[1], {'comm': '>'+c[0], 'len': len("".join(c[1:])) }) for c in b if len(c) > 1])
+
+        # UniProt Data
         if not os.path.isfile(self.db_uniprot):
-            url = self.URL_UNIPROT +'query=proteome:'+ self.proteome_id
-            if filt and filt == "sw":
-                url += '&fil=reviewed:yes'
-            url += '&format=txt'
-            logging.debug("get "+url)
+            url = self.URL_UNIPROT +'query=organism:'+ self.scientific.replace(" ",'+') +'&format=txt'
+            logging.info("get "+url+" > "+self.db_uniprot)
             urllib.request.urlretrieve(url, self.db_uniprot)
         else:
-            logging.debug('cached uniprot')
+            logging.info('cached uniprot data')
         
         # CORUM
         # download all complexes file (using the same name)
@@ -191,13 +208,13 @@ class creator:
         if not os.path.isfile(self.db_corum):
             url = self.URL_CORUM
             db_dat = self.TMP_DIR +'/'+ os.path.basename(url)
-            logging.debug("get "+url)
+            logging.info("get "+url+" > "+db_dat)
             urllib.request.urlretrieve(url, db_dat)
             zip_ref = zipfile.ZipFile(db_dat, 'r')
             zip_ref.extractall(self.TMP_DIR)
             zip_ref.close()
         else:
-            logging.debug('cached corum')
+            logging.info('cached corum')
         
         # PANTHER
         # get the list of species and extract the file name
@@ -208,163 +225,190 @@ class creator:
                 pattern = re.search(r'\s*(PTHR[^\_]*\_'+self.species+')', result, re.I | re.M)
                 if pattern:
                     url = self.URL_PANTHER + pattern[1]
-                    logging.debug("get "+url)
+                    logging.info("get "+url+" > "+self.db_panther)
                     urllib.request.urlretrieve(url, self.db_panther)
         else:
-            logging.debug('cached panther')
+            logging.info('cached panther')
     
     
     def create_qreport(self):
         '''
         Create protein report
-        '''
-        # declare output dataframe
-        df = pd.DataFrame()
-        
+        '''        
         if self.db_uniprot:
             # create reports from external data ---
             logging.info('create reports from external data...')
             corum_json = None
-            panther_txt = None
+            panther_df = None
             if os.path.isfile(self.db_corum):
                 with open(self.db_corum, 'r') as f:
                     corum_json = json.load(f)
-            logging.debug('corum done')
             if os.path.isfile(self.db_panther):
-                with open(self.db_panther, 'r') as f:
-                    panther_txt = f.read()
-            logging.debug('panther done')
+                panther_df = pd.read_csv(self.db_panther, sep="\t", dtype=str, header=None, low_memory=False)
             
-            
-            # Extract the info from the main database (UniProt), if apply
-            # create cross-references data
-            logging.info('create cross-references data from UniProtKB database...')
-            for record in SwissProt.parse( open(self.db_uniprot) ):
-                
-                # extract main info ---
-                name = record.entry_name
-                acc = record.accessions[0]
-                # accs = ";".join(record.accessions[1:])
-                pattern = re.search(r'Name=(\w*)', record.gene_name, re.I | re.M)
-                gene = pattern[1] if pattern else record.gene_name  
-                pattern = re.search(r'[RecName|SubName]: Full=([^\;|\{]*)', record.description, re.I | re.M)
-                dsc = pattern[1] if pattern else record.description
-                dclass = record.data_class
-                pattern = re.search(r'([\w|\s]*)\s+\(\w*\)', record.organism, re.I | re.M)
-                species = pattern[1] if pattern else record.organism
-                # extract isoforms IDs
-                altprod = [c for c in record.comments if 'ALTERNATIVE PRODUCTS:' in c]
-                if altprod:
-                    IsoIds = re.findall(r'IsoId=([^\;|\,]*)', altprod[0], re.I | re.M | re.DOTALL)
-                    # delete *-1 prefix from isoform Ids
-                    IsoIds = [ i.replace('-1','') for i in IsoIds ]
-                else:
-                    IsoIds = [acc]
-                # extract the comment line of fasta
-                comm = ''
-                if acc in self.db_fasta_seqio:
-                    comm = ">"+self.db_fasta_seqio[acc].description
-                    
-                
-                # create a dataframe with the Metadata information ---
-                # UniProt accesion isoform is the index
-                # ['xref_UniProt_Name','Protein','Gene','Species','Description','Comment_Line','prot_UniProt_Class']
-                df1 = pd.DataFrame(columns=self.META, data=[[name,IsoIds,gene,species,dsc,comm,dclass]])
-                df1 = df1.explode(self.XID)
-                df1.set_index(self.XID, inplace=True)
-                
-                
-                # create cross-references data ---
-                # filter by given list of terms
-                # create dictionary with all common Xreferences
-                terms_dbs = [i[0] for i in self.XTERMS]
-                rs = [x for x in record.cross_references if x[0] in terms_dbs ]
-                rcross = dict()
-                for r in rs:
-                    rcross.setdefault(r[0], []).append(r[1:])
-                
-                
-                # extract the xreference data ---
-                for terms in self.XTERMS:
-                    xdb = terms[0]
-                    xpats = terms[1]
-                    xcols,xvals = [],[]
-                    if xdb in rcross:
-                        rconts = rcross[xdb]
-                        if   xdb == "Ensembl":
-                            (xcols, xvals) = self._extract_xref_ids(rconts, xpats, acc)
-                        elif xdb == "RefSeq":
-                            (xcols, xvals) = self._extract_xref_ids(rconts, xpats, acc)
-                        elif xdb == "CCDS":
-                            (xcols, xvals) = self._extract_xref_ids(rconts, xpats, acc)
-                    else:
-                        # create empty dict with the name of colum
-                        for xc,xv in xpats:
-                            xcols.append(xc)
-                            xvals.append([np.nan])
-                        xvals = list(map(list, zip(*xvals)))
-                    # create dataframe with the Xreferece information
-                    df2 = pd.DataFrame(columns=xcols, data=xvals)
-                    if not df2.dropna().empty:
-                        # check if UniProt accession does not exit
-                        # we add the given Isoforms ids
-                        if df2[self.XID].dropna().empty:
-                            df2[self.XID] = IsoIds                        
-                        df2.set_index(self.XID, inplace=True)
-                        # join using the index which is the UniProt accession of isoform
-                        df1 = df1.join(df2, how='outer')
-                
-                
-                # create cross-references data ---
-                # filter by given list of terms
-                # create dictionary with all common Xreferences
-                terms_dbs = [i[0] for i in self.CTERMS]
-                rs = [x for x in record.cross_references if x[0] in terms_dbs ]
-                rcross = dict()
-                for r in rs:
-                    rcross.setdefault(r[0], []).append(r[1:])
+            logging.info("start the extraction of data...")
+            ddf = []
+            for rec in SwissProt.parse(open(self.db_uniprot)):
+                ddf.append( self._create_qreport(rec,corum_json,panther_df) )
+            ddf = pd.concat(ddf)
 
-                # extract the category data ---
-                for terms in self.CTERMS:
-                    xdb = terms[0]
-                    xpats = terms[1]
-                    xcols,xvals = [],[]
-                    if xdb in rcross:
-                        rconts = rcross[xdb]
-                        if xdb == "GO":
-                            (xcols, xvals) = self._extract_cat_go(rconts, xpats)
-                        elif xdb == "KEGG": # remote access
-                            (xcols, xvals) = self._extract_cat_kegg(rconts, xpats)
-                        elif xdb == "PANTHER":
-                            (xcols, xvals) = self._extract_cat_panther(panther_txt, rconts, xpats, acc)
-                        elif xdb == "Reactome":
-                            (xcols, xvals) = self._extract_cat_reactome(rconts, xpats)
-                        elif xdb == "CORUM":
-                            (xcols, xvals) = self._extract_cat_corum(corum_json, xpats, acc)
-                        elif xdb == "DrugBank":
-                            (xcols, xvals) = self._extract_cat_drugbank(rconts, xpats)
-                    else:
-                        # create empty dict with the name of colum
-                        for xc,xv in xpats:
-                            xcols.append(xc)
-                            xvals.append([np.nan])
-                        xvals = list(map(list, zip(*xvals)))
-                    # create dataframe with the Xreferece information
-                    df2 = pd.DataFrame(columns=xcols, data=xvals)
-                    if not df2.dropna().empty:
-                        # we add the given Isoforms ids
-                        dfx = pd.DataFrame(columns=[self.XID], data=IsoIds)
-                        df2 = pd.concat([df2,dfx],axis=1).ffill()
-                        df2.set_index(self.XID, inplace=True)
-                        # join using the index which is the UniProt accession of isoform
-                        df1 = df1.join(df2, how='outer')
+        return ddf
+
+
+    def _create_qreport(self, record, corum_json, panther_df):
+        '''
+        Create protein report
+        '''
+        # Efficient way to unnest (explode) multiple list columns in a pandas DataFrame
+        def _explode(df, lst_cols, fill_value=''):
+            # make sure `lst_cols` is a list
+            if lst_cols and not isinstance(lst_cols, list):
+                lst_cols = [lst_cols]
+            # all columns except `lst_cols`
+            idx_cols = df.columns.difference(lst_cols)
+        
+            # calculate lengths of lists
+            lens = df[lst_cols[0]].str.len()
+        
+            if (lens > 0).all():
+                # ALL lists in cells aren't empty
+                return pd.DataFrame({
+                    col:np.repeat(df[col].values, df[lst_cols[0]].str.len())
+                    for col in idx_cols
+                }).assign(**{col:np.concatenate(df[col].values) for col in lst_cols}) \
+                  .loc[:, df.columns]
+            else:
+                # at least one list in cells is empty
+                return pd.DataFrame({
+                    col:np.repeat(df[col].values, df[lst_cols[0]].str.len())
+                    for col in idx_cols
+                }).assign(**{col:np.concatenate(df[col].values) for col in lst_cols}) \
+                  .append(df.loc[lens==0, idx_cols]).fillna(fill_value) \
+                  .loc[:, df.columns]
+          
+        
+        # extract main info ---
+        name = record.entry_name
+        acc = record.accessions[0]
+        pattern = re.search(r'Name=(\w*)', record.gene_name, re.I | re.M)
+        gene = pattern[1] if pattern else record.gene_name  
+        pattern = re.search(r'[RecName|SubName]: Full=([^\;|\{]*)', record.description, re.I | re.M)
+        dsc = pattern[1] if pattern else record.description
+        dclass = record.data_class
+        pattern = re.search(r'([\w|\s]*)\s+\(\w*\)', record.organism, re.I | re.M)
+        species = pattern[1] if pattern else record.organism
+        # extract isoforms IDs
+        altprod = [c for c in record.comments if 'ALTERNATIVE PRODUCTS:' in c]
+        IsoIds = [acc]
+        if altprod:            
+            i = re.findall(r'IsoId=([^\;|\,]*)[\;|\,]+\s*Sequence=VSP_', altprod[0], re.I | re.M | re.DOTALL)
+            IsoIds += i
+        # for each isoform id
+        # get the comment and length of isoform
+        comms = []
+        lengths = []
+        for i in IsoIds:
+            # get the comm and len from iso_id
+            c = self.db_fasta_seqio[i]['comm'] if i in self.db_fasta_seqio else ''
+            comms.append(c)
+            l = self.db_fasta_seqio[i]['len'] if i in self.db_fasta_seqio else ''
+            lengths.append(l)
+        
+            
+        
+        # create a dataframe with the Metadata information ---
+        # UniProt accesion isoform is the index
+        # ['xref_UniProt_Name','Protein','Gene','Species','Length','Description','Comment_Line','prot_UniProt_Class']
+        df1 = pd.DataFrame(columns=self.META, data=[[name,IsoIds,gene,species,lengths,dsc,comms,dclass]])
+        df1 = _explode(df1, lst_cols=['Protein','Length','Comment_Line'])
+        df1.set_index(self.XID, inplace=True)
+        
+        # create cross-references data ---
+        # filter by given list of terms
+        # create dictionary with all common Xreferences
+        terms_dbs = [i[0] for i in self.XTERMS]
+        rs = [x for x in record.cross_references if x[0] in terms_dbs ]
+        rcross = dict()
+        for r in rs:
+            rcross.setdefault(r[0], []).append(r[1:])
+        
+        
+        # extract the xreference data ---
+        for terms in self.XTERMS:
+            xdb = terms[0]
+            xpats = terms[1]
+            xcols,xvals = [],[]
+            if xdb in rcross:
+                rconts = rcross[xdb]
+                if   xdb == "Ensembl":
+                    (xcols, xvals) = self._extract_xref_ids(rconts, xpats, acc)
+                elif xdb == "RefSeq":
+                    (xcols, xvals) = self._extract_xref_ids(rconts, xpats, acc)
+                elif xdb == "CCDS":
+                    (xcols, xvals) = self._extract_xref_ids(rconts, xpats, acc)
+            else:
+                # create empty dict with the name of colum
+                for xc,xv in xpats:
+                    xcols.append(xc)
+                    xvals.append([np.nan])
+                xvals = list(map(list, zip(*xvals)))
+            # create dataframe with the Xreferece information
+            df2 = pd.DataFrame(columns=xcols, data=xvals)
+            if not df2.dropna().empty:
+                # check if UniProt accession does not exit
+                # we add the given Isoforms ids
+                if df2[self.XID].dropna().empty:
+                    df2[self.XID] = IsoIds                        
+                df2.set_index(self.XID, inplace=True)
+                # join using the index which is the UniProt accession of isoform
+                df1 = df1.join(df2, how='outer')
+        
+        
+        # create cross-references data ---
+        # filter by given list of terms
+        # create dictionary with all common Xreferences
+        terms_dbs = [i[0] for i in self.CTERMS]
+        rs = [x for x in record.cross_references if x[0] in terms_dbs ]
+        rcross = dict()
+        for r in rs:
+            rcross.setdefault(r[0], []).append(r[1:])
+
+        # extract the category data ---
+        for terms in self.CTERMS:
+            xdb = terms[0]
+            xpats = terms[1]
+            xcols,xvals = [],[]
+            if xdb in rcross:
+                rconts = rcross[xdb]
+                if xdb == "GO":
+                    (xcols, xvals) = self._extract_cat_go(rconts, xpats)
+                elif xdb == "KEGG": # remote access
+                    (xcols, xvals) = self._extract_cat_kegg(rconts, xpats)
+                elif xdb == "PANTHER":
+                    (xcols, xvals) = self._extract_cat_panther(panther_df, rconts, xpats, acc)
+                elif xdb == "Reactome":
+                    (xcols, xvals) = self._extract_cat_reactome(rconts, xpats)
+                elif xdb == "CORUM":
+                    (xcols, xvals) = self._extract_cat_corum(corum_json, xpats, acc)
+                elif xdb == "DrugBank":
+                    (xcols, xvals) = self._extract_cat_drugbank(rconts, xpats)
+            else:
+                # create empty dict with the name of colum
+                for xc,xv in xpats:
+                    xcols.append(xc)
+                    xvals.append([np.nan])
+                xvals = list(map(list, zip(*xvals)))
+            # create dataframe with the Xreferece information
+            df2 = pd.DataFrame(columns=xcols, data=xvals)
+            if not df2.dropna().empty:
+                # we add the given Isoforms ids
+                dfx = pd.DataFrame(columns=[self.XID], data=IsoIds)
+                df2 = pd.concat([df2,dfx],axis=1).ffill()
+                df2.set_index(self.XID, inplace=True)
+                # join using the index which is the UniProt accession of isoform
+                df1 = df1.join(df2, how='outer')
                         
-                
-                # concatenate isoforms information
-                df = pd.concat([df,df1])
-                
-                
-        return df
+        return df1
     
     def _extract_xref_ids(self, rconts, xpats, acc):
         '''
@@ -444,15 +488,31 @@ class creator:
                 id = rcont[0]
                 rc = ''
                 try:
-                    record = REST.kegg_get(id).read()
-                    if record:
-                        pattern = re.search(r'DEFINITION\s*([^\n]*)', record, re.I | re.M)
-                        rc += pattern[1] if pattern else ''
-                        pattern = re.search(r'PATHWAY\s*([\w\W]*)MODULE', record, re.I | re.M)
-                        rc += "|"+re.sub(r'\s*\n\s*','|', pattern[1]) if pattern else ''
-                        rc = re.sub(r'[-|\|]*\s*$','', rc) # delete - or | at the end of string
-                        rc = f"{id}>{rc}"
-                        rcs.append(rc)
+                    id2 = id.replace(':','_')
+                    of = self.cached_dir_kegg +f"/{id2}.txt"
+                    of2 = self.cached_dir_kegg +f"/{id2}.dat"
+                    if os.path.isfile(of):
+                        with open(of, 'r') as f:
+                            rc = f.read()
+                    else:
+                        if os.path.isfile(of2):
+                            with open(of2, 'r') as f:
+                                record = f.read()
+                        else:
+                            record = REST.kegg_get(id).read()
+                            with open(of2, 'w') as f:
+                                f.write(record)
+                        if record:
+                            pattern = re.search(r'DEFINITION\s*([^\n]*)', record, re.I | re.M)
+                            rc += pattern[1] if pattern else ''
+                            pattern = re.search(r'PATHWAY\s*([\w\W]*)MODULE', record, re.I | re.M)
+                            rc += "|"+re.sub(r'\s*\n\s*','|', pattern[1]) if pattern else ''
+                            rc = re.sub(r'[-|\|]*\s*$','', rc) # delete - or | at the end of string
+                            rc = f"{id}>{rc}"
+                            with open(of, 'w') as f:
+                                f.write(rc)
+                    rcs.append(rc)
+                        
                     pass
                 except:
                     pass
@@ -468,7 +528,7 @@ class creator:
         xvals = list(map(list, zip(*xvals)))
         return (xcols, xvals)
 
-    def _extract_cat_panther(self, datatxt, rconts, xpats, acc):
+    def _extract_cat_panther(self, df, rconts, xpats, acc):
         '''
         Parse the raw database file
         '''
@@ -477,12 +537,18 @@ class creator:
         # go through all columns of xterms
         for xpat in xpats:
             xc = xpat[0] # column name
-            # xp = xpat[1] # pattern
             # extract the information from the given UniProt accession
-            rcs = ''
-            if datatxt:
-                pattern = re.search(rf"UniProtKB={acc}\t*([^\t]*)\t*([^\t]*)", datatxt, re.I | re.M)
-                rcs += f"{pattern[1]}>{pattern[2]}" if pattern else ''
+            try:
+                x = df[df[1] == acc][[3,5]].values.tolist()[0]
+                rcs = f"{x[0]}>{x[1]}"
+                pass
+            except Exception:
+                rcs = ''
+                pass
+            
+            # if datatxt:
+            #     pattern = re.search(rf"UniProtKB={acc}\t*([^\t]*)\t*([^\t]*)", datatxt, re.I | re.M)
+            #     rcs += f"{pattern[1]}>{pattern[2]}" if pattern else ''
             # otherwise, we use the information from given records
             if rcs == '':
                 rcs = ";".join([x[0] for x in rconts])